Refactor configs package - Remove getConfig and replace it with getLocaleConfig - Introduce dictionaries/index.ts and Dictionary type
