+++ packages/configs/src/dictionaries/index.ts
@@ -1,0 +1,43 @@
+import { Dictionary, Locale } from '@scrabble-solver/types';
+
+export const fise2: Dictionary = {
+  description: 'github.com/kamilmielnik/fise-2',
+  locale: Locale.ES_ES,
+  name: 'FISE-2',
+};
+
+export const french: Dictionary = {
+  description: 'github.com/hbenbel/French-Dictionary',
+  locale: Locale.FR_FR,
+  name: 'Français',
+};
+
+export const german: Dictionary = {
+  description: 'github.com/hippler/german-wordlist',
+  locale: Locale.DE_DE,
+  name: 'Deutsch',
+};
+
+export const persian: Dictionary = {
+  description: 'github.com/MansourM/persian-to-persian-dictionary',
+  locale: Locale.FA_IR,
+  name: 'فارسی',
+};
+
+export const sjp: Dictionary = {
+  description: 'Słownik SJP.PL',
+  locale: Locale.PL_PL,
+  name: 'SJP.PL',
+};
+
+export const sowpods: Dictionary = {
+  description: 'Collins Scrabble Words',
+  locale: Locale.EN_GB,
+  name: 'SOWPODS',
+};
+
+export const twl06: Dictionary = {
+  description: 'NASPA Word List',
+  locale: Locale.EN_US,
+  name: 'TWL06',
+};
