+++ packages/configs/src/games/index.ts
@@ -1,0 +1,2 @@
+export { default as literaki } from './literaki';
+export { default as scrabble } from './scrabble';
