+++ packages/configs/src/index.ts
@@ -1,5 +1,4 @@
+export * from './dictionaries';
+export * from './games';
 export { default as getConfig } from './getConfig';
-export { default as getLocaleConfig } from './getLocaleConfig';
 export { default as isConfigId } from './isConfigId';
-export { default as literaki } from './literaki';
-export { default as scrabble } from './scrabble';
