+++ packages/configs/src/isConfigId.ts
@@ -1,7 +1,6 @@
-import literaki from './literaki';
-import scrabble from './scrabble';
+import * as games from './games';
 
-const configs = [literaki, scrabble];
+const configs = Object.values(games).flat();
 
 const isConfigId = (configId: unknown): boolean => {
   return configs.some(({ id }) => id === configId);
