--- packages/configs/src/literaki/index.ts
@@ -1,1 +1,0 @@
-export { default } from './literaki';
