+++ packages/configs/src/locales/enUs.ts
@@ -1,0 +1,73 @@
+import { Config, Locale } from '@scrabble-solver/types';
+
+import { literaki, scrabble } from '../games';
+
+const enUs: Config[] = [
+  new Config({
+    ...scrabble,
+    locale: Locale.EN_US,
+    tiles: [
+      { character: 'a', count: 9, score: 1 },
+      { character: 'b', count: 2, score: 3 },
+      { character: 'c', count: 2, score: 3 },
+      { character: 'd', count: 4, score: 2 },
+      { character: 'e', count: 12, score: 1 },
+      { character: 'f', count: 2, score: 4 },
+      { character: 'g', count: 3, score: 2 },
+      { character: 'h', count: 2, score: 4 },
+      { character: 'i', count: 9, score: 1 },
+      { character: 'j', count: 1, score: 8 },
+      { character: 'k', count: 1, score: 5 },
+      { character: 'l', count: 4, score: 1 },
+      { character: 'm', count: 2, score: 3 },
+      { character: 'n', count: 6, score: 1 },
+      { character: 'o', count: 8, score: 1 },
+      { character: 'p', count: 2, score: 3 },
+      { character: 'q', count: 1, score: 10 },
+      { character: 'r', count: 6, score: 1 },
+      { character: 's', count: 4, score: 1 },
+      { character: 't', count: 6, score: 1 },
+      { character: 'u', count: 4, score: 1 },
+      { character: 'v', count: 2, score: 4 },
+      { character: 'w', count: 2, score: 4 },
+      { character: 'x', count: 1, score: 8 },
+      { character: 'y', count: 2, score: 4 },
+      { character: 'z', count: 1, score: 10 },
+    ],
+  }),
+  new Config({
+    ...literaki,
+    locale: Locale.EN_US,
+    name: 'Literaxx',
+    tiles: [
+      { character: 'a', count: 9, score: 1 },
+      { character: 'b', count: 2, score: 3 },
+      { character: 'c', count: 2, score: 3 },
+      { character: 'd', count: 4, score: 2 },
+      { character: 'e', count: 12, score: 1 },
+      { character: 'f', count: 2, score: 4 },
+      { character: 'g', count: 3, score: 2 },
+      { character: 'h', count: 2, score: 4 },
+      { character: 'i', count: 9, score: 1 },
+      { character: 'j', count: 1, score: 8 },
+      { character: 'k', count: 1, score: 5 },
+      { character: 'l', count: 4, score: 1 },
+      { character: 'm', count: 2, score: 3 },
+      { character: 'n', count: 6, score: 1 },
+      { character: 'o', count: 8, score: 1 },
+      { character: 'p', count: 2, score: 3 },
+      { character: 'q', count: 1, score: 10 },
+      { character: 'r', count: 6, score: 1 },
+      { character: 's', count: 4, score: 1 },
+      { character: 't', count: 6, score: 1 },
+      { character: 'u', count: 4, score: 1 },
+      { character: 'v', count: 2, score: 4 },
+      { character: 'w', count: 2, score: 4 },
+      { character: 'x', count: 1, score: 8 },
+      { character: 'y', count: 2, score: 4 },
+      { character: 'z', count: 1, score: 10 },
+    ],
+  }),
+];
+
+export default enUs;
