+++ packages/configs/src/locales/index.ts
@@ -1,0 +1,7 @@
+export { default as deDe } from './deDe';
+export { default as enGb } from './enGb';
+export { default as enUs } from './enUs';
+export { default as esEs } from './esEs';
+export { default as frFr } from './frFr';
+export { default as faIr } from './faIr';
+export { default as plPl } from './plPl';
