--- packages/configs/src/scrabble/index.ts
@@ -1,1 +1,0 @@
-export { default } from './scrabble';
