--- packages/configs/src/scrabble/tiles/index.ts
@@ -1,6 +1,0 @@
-export { default as tilesEn } from './tilesEn';
-export { default as tilesEs } from './tilesEs';
-export { default as tilesFa } from './tilesFa';
-export { default as tilesFr } from './tilesFr';
-export { default as tilesPl } from './tilesPl';
-export { default as tilesDe } from './tilesDe';
