+++ packages/scrabble-solver/src/state/selectors.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-lines */
 
 import { createSelector } from '@reduxjs/toolkit';
-import { getLocaleConfig } from '@scrabble-solver/configs';
+import { getConfig } from '@scrabble-solver/configs';
 import { BLANK } from '@scrabble-solver/constants';
 import { Cell, Config, isError, Tile } from '@scrabble-solver/types';
 
@@ -69,7 +69,7 @@ export const selectInputMode = createSelector([selectSettingsRoot], (settings) =
 
 export const selectConfigId = createSelector([selectSettingsRoot], (settings) => settings.configId);
 
-export const selectConfig = createSelector([selectConfigId, selectLocale], getLocaleConfig);
+export const selectConfig = createSelector([selectConfigId, selectLocale], getConfig);
 
 export const selectFilteredCells = selectCellFilterRoot;
