+++ packages/scrabble-solver/src/state/slices/boardInitialState.ts
@@ -1,4 +1,4 @@
-import { getLocaleConfig } from '@scrabble-solver/configs';
+import { getConfig } from '@scrabble-solver/configs';
 import { Board } from '@scrabble-solver/types';
 
 import localStorage from '../localStorage';
@@ -8,7 +8,7 @@ import settingsInitialState from './settingsInitialState';
 export type BoardState = Board;
 
 const { configId, locale } = settingsInitialState;
-const { boardHeight, boardWidth } = getLocaleConfig(configId, locale);
+const { boardHeight, boardWidth } = getConfig(configId, locale);
 export const boardDefaultState = Board.create(boardWidth, boardHeight);
 
 const boardInitialState: BoardState = localStorage.getBoard() || boardDefaultState;
