+++ packages/types/src/Dictionary.ts
@@ -1,0 +1,9 @@
+import Locale from './Locale';
+
+interface Dictionary {
+  description: string;
+  locale: Locale;
+  name: string;
+}
+
+export default Dictionary;
