+++ packages/types/src/index.ts
@@ -10,6 +10,7 @@ export { default as Collision } from './Collision';
 export { default as CollisionJson } from './CollisionJson';
 export { default as Config } from './Config';
 export { default as ConfigJson } from './ConfigJson';
+export { default as Dictionary } from './Dictionary';
 export { default as FinalPattern } from './FinalPattern';
 export { default as HorizontalPattern } from './HorizontalPattern';
 export { default as isError } from './isError';
